requirements.txt
dash==2.3.1
dash_bootstrap_components==1.1.0
index==0.2
matplotlib==3.5.1
numpy==1.22.3
Pillow==9.1.0
scikit_image==0.19.2
skimage==0.0
tensorflow==2.6.0
tqdm==4.64.0
waitress==2.1.1
